fix: Fix file paths in Category components
- Category 컴포넌트의 파일 경로를 수정했습니다.
- 파일 경로를 올바르게 변경하여 코드를 업데이트했습니다.
